requirements.txt
Flask
opencv-python
pytesseract
pdf2image
torch
torchvision
numpy
Pillow
pandas
requests